fix: Restore old feeds in addition to still providing new AP feeds
The feeds still need to go into the old bucket because JPC needs to read
them. Output these in addition to the new analytical platform feeds.
