+++ app/lib/cloud_data/reports_feed.rb
@@ -1,7 +1,5 @@
 module CloudData
   class ReportsFeed
-    DATABASE_NAME = 'feeds_report'.freeze
-
     def initialize(bucket_name = ENV['S3_REPORTING_BUCKET_NAME'])
       client = Aws::S3::Client.new(
         access_key_id: ENV['S3_REPORTING_ACCESS_KEY_ID'],
@@ -11,26 +9,19 @@ module CloudData
       @bucket = @s3.bucket(bucket_name)
     end
 
-    def write(content, table, report_date = Time.zone.yesterday)
-      path = full_path(report_date, table)
+    def write(content, obj_name, report_date = Time.zone.yesterday)
+      folder_name = report_date.strftime('%Y/%m/%d')
+      filename = report_date.strftime('%Y-%m-%d')
 
-      obj = bucket.object(path)
-      obj.put(acl: 'bucket-owner-full-control', body: content, server_side_encryption: 'AES256')
+      full_name = "#{folder_name}/#{filename}-#{obj_name}"
+      obj = bucket.object(full_name)
+      obj.put(body: content)
 
-      path
+      full_name
     end
 
   private
 
     attr_reader :bucket
-
-    def full_path(report_date, table)
-      "#{ENV.fetch('S3_REPORTING_PROJECT_PATH')}/" \
-      'data/' \
-      "database_name=#{DATABASE_NAME}/" \
-      "table_name=#{table}/" \
-      "extraction_timestamp=#{report_date.strftime('%Y%m%d%H%M%SZ')}/" \
-      "#{report_date.strftime('%Y-%m-%d')}-#{table}.jsonl"
-    end
   end
 end
